refactor: Restructure backend into modular classification and task domains.
