+++ backend/src/main/java/com/swipelab/classification/application/ConsensusService.java
@@ -1,0 +1,5 @@
+package com.swipelab.classification.application;
+
+public class ConsensusService {
+    
+}
