+++ backend/src/main/java/com/swipelab/classification/application/TaskDistributionService.java
@@ -1,8 +1,8 @@
-package com.swipelab.service.classification;
+package com.swipelab.classification.application;
 
-import com.swipelab.model.entity.Image;
-import com.swipelab.repository.ClassificationRepository;
-import com.swipelab.repository.ImageRepository;
+import com.swipelab.classification.domain.Image;
+import com.swipelab.classification.infrastructure.ClassificationRepository;
+import com.swipelab.classification.infrastructure.ImageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
