+++ backend/src/main/java/com/swipelab/classification/infrastructure/ClassificationRepository.java
@@ -1,6 +1,6 @@
-package com.swipelab.repository;
+package com.swipelab.classification.infrastructure;
 
-import com.swipelab.model.entity.Classification;
+import com.swipelab.classification.domain.Classification;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
