+++ backend/src/main/java/com/swipelab/classification/infrastructure/ImageRepository.java
@@ -1,6 +1,6 @@
-package com.swipelab.repository;
+package com.swipelab.classification.infrastructure;
 
-import com.swipelab.model.entity.Image;
+import com.swipelab.classification.domain.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
