+++ backend/src/main/java/com/swipelab/service/analytics/AnalyticsService.java
@@ -3,13 +3,13 @@ package com.swipelab.service.analytics;
 import com.swipelab.dto.response.TaskAnalyticsResponse;
 import com.swipelab.dto.response.UserPerformanceResponse;
 import com.swipelab.exception.ResourceNotFoundException;
-import com.swipelab.model.entity.Classification;
-import com.swipelab.model.entity.Image;
-import com.swipelab.model.entity.Task;
+import com.swipelab.classification.domain.Classification;
+import com.swipelab.classification.domain.Image;
+import com.swipelab.tasks.domain.Task;
 import com.swipelab.model.entity.User;
-import com.swipelab.repository.ClassificationRepository;
-import com.swipelab.repository.ImageRepository;
-import com.swipelab.repository.TaskRepository;
+import com.swipelab.classification.infrastructure.ClassificationRepository;
+import com.swipelab.classification.infrastructure.ImageRepository;
+import com.swipelab.tasks.infrastructure.TaskRepository;
 import com.swipelab.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
