--- backend/src/main/java/com/swipelab/service/classification/ValidationService.java
@@ -1,5 +1,0 @@
-package com.swipelab.service.classification;
-
-public class ValidationService {
-    
-}
